refactor: remove code duplication of several `Conversation` methods\
Namely, `get_device_identities()`, `get_user_identities()`, and
`e2ei_conversation_state()`
