+++ crypto/src/mls/conversation/conversation_guard/e2e_identity/mod.rs
@@ -1,6 +1,4 @@
-mod identity;
 mod rotate;
 
 use super::ConversationGuard;
-use super::Error;
 use super::Result;
